+++ x-pack/plugins/alerting/common/alert_schema/field_maps/mapping_from_field_map.test.ts
@@ -255,7 +255,7 @@ describe('mappingFromFieldMap', () => {
                       type: 'keyword',
                     },
                     revision: {
-                      type: 'integer',
+                      type: 'long',
                     },
                     rule_type_id: {
                       type: 'keyword',
